# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HelpMotivateMe.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/buddies/invite": {
      "post": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteBuddyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteBuddyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InviteBuddyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/buddies/{id}": {
      "delete": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/buddies/leave/{ownerUserId}": {
      "delete": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "ownerUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/buddies/{targetUserId}/today": {
      "get": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyTodayViewResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyTodayViewResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyTodayViewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/buddies/{targetUserId}/journal": {
      "get": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuddyJournalEntryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuddyJournalEntryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuddyJournalEntryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuddyJournalEntryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuddyJournalEntryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuddyJournalEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalEntryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalEntryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/buddies/{targetUserId}/journal/{entryId}/images": {
      "post": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalImageResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalImageResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalImageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/buddies/{targetUserId}/journal/{entryId}/reactions": {
      "post": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBuddyJournalReactionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBuddyJournalReactionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddBuddyJournalReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalReactionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalReactionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuddyJournalReactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/buddies/{targetUserId}/journal/{entryId}/reactions/{reactionId}": {
      "delete": {
        "tags": [
          "AccountabilityBuddy"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddBuddyJournalReactionRequest": {
        "required": [
          "emoji"
        ],
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string"
          }
        }
      },
...
      "BuddyJournalEntryResponse": {
        "required": [
          "id",
          "title",
          "description",
          "entryDate",
          "authorUserId",
          "authorDisplayName",
          "images",
          "reactions",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "entryDate": {
            "type": "string"
          },
          "authorUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "authorDisplayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuddyJournalImageResponse"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BuddyJournalReactionResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BuddyJournalImageResponse": {
        "required": [
          "id",
          "fileName",
          "url",
          "sortOrder"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BuddyJournalReactionResponse": {
        "required": [
          "id",
          "emoji",
          "userId",
          "userDisplayName",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "emoji": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userDisplayName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BuddyResponse": {
        "required": [
          "id",
          "buddyUserId",
          "buddyEmail",
          "buddyDisplayName",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "buddyUserId": {
            "type": "string",
            "format": "uuid"
          },
          "buddyEmail": {
            "type": "string"
          },
          "buddyDisplayName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BuddyTodayViewResponse": {
        "required": [
          "userId",
          "userDisplayName",
          "date",
          "habitStacks",
          "upcomingTasks",
          "completedTasks",
          "identityFeedback"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userDisplayName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "habitStacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayHabitStackResponse"
            }
          },
          "upcomingTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayTaskResponse"
            }
          },
          "completedTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayTaskResponse"
            }
          },
          "identityFeedback": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayIdentityFeedbackResponse"
            }
          }
        }
      },
...
      "CreateBuddyJournalEntryRequest": {
        "required": [
          "title",
          "description",
          "entryDate"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "entryDate": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "InviteBuddyRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "TodayHabitStackItemResponse": {
        "required": [
          "id",
          "habitDescription",
          "isCompletedToday",
          "currentStreak"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "habitDescription": {
            "type": "string"
          },
          "isCompletedToday": {
            "type": "boolean"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TodayHabitStackResponse": {
        "required": [
          "id",
          "name",
          "triggerCue",
          "identityId",
          "identityName",
          "identityColor",
          "identityIcon",
          "items",
          "completedCount",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "triggerCue": {
            "type": [
              "null",
              "string"
            ]
          },
          "identityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "identityName": {
            "type": [
              "null",
              "string"
            ]
          },
          "identityColor": {
            "type": [
              "null",
              "string"
            ]
          },
          "identityIcon": {
            "type": [
              "null",
              "string"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayHabitStackItemResponse"
            }
          },
          "completedCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TodayIdentityFeedbackResponse": {
        "required": [
          "id",
          "name",
          "color",
          "icon",
          "totalVotes",
          "habitVotes",
          "stackBonusVotes",
          "taskVotes",
          "proofVotes",
          "reinforcementMessage"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "icon": {
            "type": [
              "null",
              "string"
            ]
          },
          "totalVotes": {
            "type": "integer",
            "format": "int32"
          },
          "habitVotes": {
            "type": "integer",
            "format": "int32"
          },
          "stackBonusVotes": {
            "type": "integer",
            "format": "int32"
          },
          "taskVotes": {
            "type": "integer",
            "format": "int32"
          },
          "proofVotes": {
            "type": "integer",
            "format": "int32"
          },
          "reinforcementMessage": {
            "type": "string"
          }
        }
      },
      "TodayTaskResponse": {
        "required": [
          "id",
          "title",
          "description",
          "goalId",
          "goalTitle",
          "identityId",
          "identityName",
          "identityIcon",
          "identityColor",
          "dueDate",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "goalId": {
            "type": "string",
            "format": "uuid"
          },
          "goalTitle": {
            "type": "string"
          },
          "identityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "identityName": {
            "type": [
              "null",
              "string"
            ]
          },
          "identityIcon": {
            "type": [
              "null",
              "string"
            ]
          },
          "identityColor": {
            "type": [
              "null",
              "string"
            ]
          },
          "dueDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}